Critical JWSTreader changes
Note that the JWST module now saves non_static attributes as normal arrays (not arrays of arrays). This might cause some problems with other modules that have been written recently, namely if an attribute was read within a module using get_attribute, previously the get_attribute() function had to be combined with [0], which is now not necessary anymore. (The changes took effect last commit.